media app/requirements.txt
requests
PySide6
python-dotenv